+++ database/migrations/2023_04_10_135856_create_detalleingresos_table.php
@@ -16,10 +16,10 @@ return new class extends Migration
             $table->unsignedBigInteger('ingreso_id');
             $table->unsignedBigInteger('product_id');
             $table->integer('cantidad');
-            $table->integer('preciounitario');
-            $table->integer('preciounitariomo');
-            $table->integer('servicio');
-            $table->integer('preciofinal');
+            $table->double('preciounitario');
+            $table->double('preciounitariomo');
+            $table->double('servicio');
+            $table->double('preciofinal');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->foreign('ingreso_id')->references('id')->on('ingresos')->onDelete('cascade');
             $table->timestamps();
